+++ src/main/java/com/github/addshore/facebook/data/image/exif/MainOptions.java
@@ -5,15 +5,27 @@ public class MainOptions {
     private final Boolean debug;
     private final Boolean dry;
     private final Boolean overwriteOriginals;
+    private final Boolean copyToNewFolders;
+    private final Boolean processPosts;
+    private final Boolean processMessages;
+    private final Boolean processFiles;
 
     public MainOptions(
             Boolean debug,
             Boolean dry,
-            Boolean overwriteOriginals
+            Boolean overwriteOriginals,
+            Boolean copyToNewFolders,
+            Boolean processPosts,
+            Boolean processMessages,
+            Boolean processFiles
     ) {
         this.debug = debug;
         this.dry = dry;
         this.overwriteOriginals = overwriteOriginals;
+        this.copyToNewFolders = copyToNewFolders;
+        this.processPosts = processPosts;
+        this.processMessages = processMessages;
+        this.processFiles = processFiles;
     }
 
     public Boolean isDryMode() {
@@ -26,6 +38,22 @@ public class MainOptions {
 
     public Boolean shouldOverwriteOriginals() {
         return overwriteOriginals;
+    }
+
+    public Boolean shouldCopyToNewFolders() {
+        return copyToNewFolders;
+    }
+
+    public Boolean shouldProcessPosts() {
+        return processPosts;
+    }
+
+    public Boolean shouldProcessMessages() {
+        return processMessages;
+    }
+
+    public Boolean shouldProcessFiles() {
+        return processFiles;
     }
 
 }